part of adjust structure of oltp
