+++ hugegraph-core/src/main/java/com/baidu/hugegraph/traversal/algorithm/strategy/ConcurrentTraverseStrategy.java
@@ -28,6 +28,7 @@ import java.util.function.BiConsumer;
 
 import com.baidu.hugegraph.HugeGraph;
 import com.baidu.hugegraph.backend.id.Id;
+import com.baidu.hugegraph.traversal.algorithm.ConcurrentMultiValuedMap;
 import com.baidu.hugegraph.traversal.algorithm.steps.EdgeStep;
 import com.baidu.hugegraph.traversal.algorithm.OltpTraverser;
 
@@ -40,7 +41,7 @@ public class ConcurrentTraverseStrategy extends OltpTraverser
 
     @Override
     public Map<Id, List<Node>> newMultiValueMap() {
-        return new OltpTraverser.ConcurrentMultiValuedMap<>();
+        return new ConcurrentMultiValuedMap<>();
     }
 
     @Override
